+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Bad hyperlink for release in CHANGELOG.
-
+- Missing "year" variable from root of output climate data file ([PR #6](https://github.com/fact-sealevel/fair-temperature/pull/6), [@brews](https://github.com/brews)).
 
 
 ## [0.2.0] - 2025-10-17